requirements.txt
Flask>=2.0.0
requests>=2.28.0
scikit-learn>=1.0.0
joblib==1.3.2
gunicorn==20.1.0
python-dotenv>=0.19.0
flask-healthz==1.0.1
prometheus-client>=0.16.0
flask-login==0.6.2
python-json-logger==2.0.7
Flask-SocketIO>=5.1.1
eventlet>=0.33.0
python-engineio==4.3.4
python-socketio>=5.4.0
pytest==7.4.0
pytest-cov==4.1.0
numpy>=1.21.0
pandas>=1.3.0
flask-cors>=3.0.10
pyOpenSSL>=20.0.1
APScheduler>=3.9.1
tenacity>=8.0.1
pyotp>=2.8.0
